+++ src/views/system/menu.vue
@@ -13,7 +13,7 @@
               size="small"
               icon="PlusIcon"
               @click="onAddItem"
-              >添加
+            >添加
             </el-button>
           </template>
         </TableConfig>
@@ -29,14 +29,31 @@
           row-key="menuUrl"
           :tree-props="{ children: 'children' }"
         >
-          <el-table-column align="center" label="序号" fixed="left" width="150">
+          <el-table-column
+            align="center"
+            label="序号"
+            fixed="left"
+            width="150"
+          >
             <template v-slot="scope">
               {{ scope.$index + 1 }}
             </template>
           </el-table-column>
-          <el-table-column align="center" label="菜单名称" prop="menuName" />
-          <el-table-column align="center" label="菜单地址" prop="menuUrl" />
-          <el-table-column align="center" label="图标" prop="menuUrl">
+          <el-table-column
+            align="center"
+            label="菜单名称"
+            prop="menuName"
+          />
+          <el-table-column
+            align="center"
+            label="菜单地址"
+            prop="menuUrl"
+          />
+          <el-table-column
+            align="center"
+            label="图标"
+            prop="menuUrl"
+          >
             <template #default="scope">
               <el-icon
                 v-if="scope.row.icon"
@@ -49,25 +66,33 @@
               <div v-else>--</div>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="是否缓存">
+          <el-table-column
+            align="center"
+            label="是否缓存"
+          >
             <template #default="scope">
               <el-tag
                 :type="scope.row.cacheable ? 'primary' : 'danger'"
                 size="small"
-                >{{ scope.row.cacheable ? "是" : "否" }}</el-tag
-              >
+              >{{ scope.row.cacheable ? "是" : "否" }}</el-tag>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="是否固定">
+          <el-table-column
+            align="center"
+            label="是否固定"
+          >
             <template #default="scope">
               <el-tag
                 :type="scope.row.affix ? 'primary' : 'danger'"
                 size="small"
-                >{{ scope.row.affix ? "是" : "否" }}
+              >{{ scope.row.affix ? "是" : "否" }}
               </el-tag>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="badge提示">
+          <el-table-column
+            align="center"
+            label="badge提示"
+          >
             <template #default="scope">
               <div class="menu-badge__wrapper">
                 <el-badge
@@ -82,22 +107,25 @@
               </div>
             </template>
           </el-table-column>
-          <el-table-column align="center" label="操作" fixed="right">
+          <el-table-column
+            align="center"
+            :width="200"
+            label="操作"
+            fixed="right"
+          >
             <template v-slot="scope">
               <el-button
                 plain
                 type="primary"
                 size="small"
                 @click="onUpdateItem(scope.row)"
-                >编辑</el-button
-              >
+              >编辑</el-button>
               <el-button
                 plain
                 type="danger"
                 size="small"
                 @click="onDeleteItem(scope.row)"
-                >删除</el-button
-              >
+              >删除</el-button>
             </template>
           </el-table-column>
         </el-table>
@@ -123,14 +151,20 @@
               }"
             />
           </el-form-item>
-          <el-form-item label="菜单名称" prop="name">
+          <el-form-item
+            label="菜单名称"
+            prop="name"
+          >
             <el-input
               v-model="menuModel.name"
               size="small"
               placeholder="请输入菜单名称"
             />
           </el-form-item>
-          <el-form-item label="菜单地址" prop="path">
+          <el-form-item
+            label="菜单地址"
+            prop="path"
+          >
             <el-input
               v-model="menuModel.path"
               size="small"
@@ -147,7 +181,10 @@
             </el-input>
           </el-form-item>
           <el-form-item label="badge提示">
-            <el-radio-group v-model="menuModel.badge" size="small">
+            <el-radio-group
+              v-model="menuModel.badge"
+              size="small"
+            >
               <el-radio-button label="">无</el-radio-button>
               <el-radio-button label="dot">圆点</el-radio-button>
               <el-radio-button label="new">new</el-radio-button>